requirements.txt
indic-nlp-library>=0.92
torch==2.0.1
transformers==4.41.2
tokenizers==0.19.1
datasets==2.18.0
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.2
accelerate==0.30.1
sentencepiece==0.2.0
huggingface-hub==0.23.0
scikit-learn==1.4.2
